+++ src/Controller/AppController.php
@@ -34,8 +34,5 @@ abstract class AppController extends Controller
     public function initialize(): void
     {
         parent::initialize();
-
-        $this->loadComponent('RequestHandler');
-        $this->loadComponent('MeTools.Flash');
     }
 }